Move Class and Entity related logic to separate files
